Create main.py
Imported random from Python to generate random numbers. Defined a name, question, answer, and random_number variable. Added if, elif, else conditionals to check the random number and output an answer based on the random number. Also added a conditional to check if the name variable is empty or not, and if it is to bypass the name output and just display the question.